split formatting, add --options, --noindex, fixed rsi 0
