+++ libs/metrics/market_composite_index.py
@@ -55,7 +55,10 @@ def composite_index(data: pd.DataFrame, sectors: list, plot_output=True):
     return composite2 
 
 
-def market_composite_index(period='1y', properties=None):
+def market_composite_index(config: dict):
+    period = config['period']
+    properties = config['properties']
+
     """ Validate each index key is set to True in the --core file """
     if properties is not None:
         if 'Indexes' in properties.keys():
