+++ libs/utils/__init__.py
@@ -1,9 +1,12 @@
+from .startup import start_header
+from .file_io import configure_temp_dir, remove_temp_dir, create_sub_temp_dir, windows_compatible_file_parse
+from .data import download_data, data_nan_fix
+
 from .formatting import name_parser, index_extractor, fund_list_extractor, index_appender
 from .formatting import dates_extractor_list, date_extractor, get_daterange
-from .formatting import configure_temp_dir, remove_temp_dir, create_sub_temp_dir, windows_compatible_file_parse
-from .formatting import start_header, download_data, data_nan_fix
 
 from .plotting import dual_plotting, generic_plotting, histogram, bar_chart, specialty_plotting, shape_plotting
 from .plotting import candlestick
+
 from .nasit import nasit_oscillator_score, nasit_oscillator_signal, nasit_cluster_signal, nasit_cluster_score
 from .progress_bar import ProgressBar
