+++ technical_analysis.py
@@ -14,7 +14,7 @@
 """
 
 ################################
-_VERSION_ = '0.1.12'
+_VERSION_ = '0.1.13'
 _DATE_REVISION_ = '2019-08-16'
 ################################
 
@@ -47,101 +47,109 @@ from libs.tools import get_maxima_minima, get_trendlines
 
 ####################################################################
 ####################################################################
-tickers, ticker_print, period, interval, properties = start_header(update_release=_DATE_REVISION_, version=_VERSION_)
+config = start_header(update_release=_DATE_REVISION_, version=_VERSION_)
 PROCESS_STEPS = 13
 
-tickers = index_appender(tickers)
-sp500_index = index_extractor(tickers)
+if config['state'] != 'halt':
 
-# Temporary directories to save graphs as images, etc.
-remove_temp_dir()
-configure_temp_dir()
+    if config['state'] != 'run_no_index':
+        config['tickers'] = index_appender(config['tickers'])
+        sp500_index = index_extractor(config['tickers'])
 
-data = download_data(period=period, interval=interval, ticker_print=ticker_print, tickers=tickers)
-    
-funds = fund_list_extractor(data)
+    # Temporary directories to save graphs as images, etc.
+    remove_temp_dir()
+    configure_temp_dir()
 
-# Start of automated process
-analysis = {}
+    data = download_data(config=config)
+        
+    funds = fund_list_extractor(data, config=config)
 
-for fund_name in funds:
-    
-    print(f"~~{fund_name}~~")
-    create_sub_temp_dir(fund_name)
-    analysis[fund_name] = {}
+    # Start of automated process
+    analysis = {}
 
-    p = ProgressBar(PROCESS_STEPS, name=fund_name)
-    p.start()
+    for fund_name in funds:
+        
+        print(f"~~{fund_name}~~")
+        create_sub_temp_dir(fund_name)
+        analysis[fund_name] = {}
 
-    fund = data[fund_name]
-    fund = data_nan_fix(fund)
-    p.uptick()
+        p = ProgressBar(PROCESS_STEPS, name=fund_name)
+        p.start()
 
-    start = date_extractor(fund.index[0], _format='str')
-    end = date_extractor(fund.index[len(fund['Close'])-1], _format='str')
+        if len(funds) > 1:
+            fund = data[fund_name]
+        else:
+            fund = data
 
-    analysis[fund_name]['dates_covered'] = {'start': str(start), 'end': str(end)} 
-    analysis[fund_name]['name'] = fund_name
+        fund = data_nan_fix(fund)
+        p.uptick()
 
-    chart, dat = cluster_oscs(fund, function='all', filter_thresh=3, name=fund_name, plot_output=False)
-    analysis[fund_name]['clustered_osc'] = dat
-    p.uptick()
+        start = date_extractor(fund.index[0], _format='str')
+        end = date_extractor(fund.index[len(fund['Close'])-1], _format='str')
 
-    on_balance_volume(fund, plot_output=False, name=fund_name)
-    p.uptick()
+        analysis[fund_name]['dates_covered'] = {'start': str(start), 'end': str(end)} 
+        analysis[fund_name]['name'] = fund_name
 
-    triple_moving_average(fund, plot_output=False, name=fund_name)
-    p.uptick()
+        chart, dat = cluster_oscs(fund, function='all', filter_thresh=3, name=fund_name, plot_output=False)
+        analysis[fund_name]['clustered_osc'] = dat
+        p.uptick()
 
-    moving_average_swing_trade(fund, plot_output=False, name=fund_name)
-    p.uptick()
+        on_balance_volume(fund, plot_output=False, name=fund_name)
+        p.uptick()
 
-    analysis[fund_name]['macd'] = mov_avg_convergence_divergence(fund, plot_output=False, name=fund_name)
-    p.uptick()
+        triple_moving_average(fund, plot_output=False, name=fund_name)
+        p.uptick()
 
-    analysis[fund_name]['relative_strength'] = relative_strength(fund_name, fund_name, tickers=data, sector='', plot_output=False)
-    p.uptick()
+        moving_average_swing_trade(fund, plot_output=False, name=fund_name)
+        p.uptick()
 
-    # Support and Resistance Analysis
-    analysis[fund_name]['support_resistance'] = find_resistance_support_lines(fund, name=fund_name, plot_output=False)
-    p.uptick()
+        analysis[fund_name]['macd'] = mov_avg_convergence_divergence(fund, plot_output=False, name=fund_name)
+        p.uptick()
 
-    # Feature Detection Block
-    shapes = []
-    analysis[fund_name]['features'] = {}
+        if config['state'] != 'run_no_index':
+            analysis[fund_name]['relative_strength'] = relative_strength(fund_name, fund_name, tickers=data, sector='', plot_output=False)
+        p.uptick()
 
-    hs2, ma, shapes = feature_head_and_shoulders(fund, FILTER_SIZE=2, name=fund_name, shapes=shapes)
-    analysis[fund_name]['features']['head_shoulders_2'] = hs2
-    p.uptick()
+        # Support and Resistance Analysis
+        analysis[fund_name]['support_resistance'] = find_resistance_support_lines(fund, name=fund_name, plot_output=False)
+        p.uptick()
 
-    hs, ma, shapes = feature_head_and_shoulders(fund, FILTER_SIZE=4, name=fund_name, shapes=shapes)
-    analysis[fund_name]['features']['head_shoulders_4'] = hs
-    p.uptick()
+        # Feature Detection Block
+        shapes = []
+        analysis[fund_name]['features'] = {}
 
-    hs3, ma, shapes = feature_head_and_shoulders(fund, FILTER_SIZE=8, name=fund_name, shapes=shapes)
-    analysis[fund_name]['features']['head_shoulders_8'] = hs3
-    p.uptick()
+        hs2, ma, shapes = feature_head_and_shoulders(fund, FILTER_SIZE=2, name=fund_name, shapes=shapes)
+        analysis[fund_name]['features']['head_shoulders_2'] = hs2
+        p.uptick()
 
-    feature_plotter(fund, shapes, name=fund_name, feature='head_and_shoulders')
-    p.uptick()
+        hs, ma, shapes = feature_head_and_shoulders(fund, FILTER_SIZE=4, name=fund_name, shapes=shapes)
+        analysis[fund_name]['features']['head_shoulders_4'] = hs
+        p.uptick()
 
-    filename = f"{fund_name}/candlestick_{fund_name}"
-    candlestick(fund, title=fund_name, filename=filename, saveFig=True)
-    p.uptick()
+        hs3, ma, shapes = feature_head_and_shoulders(fund, FILTER_SIZE=8, name=fund_name, shapes=shapes)
+        analysis[fund_name]['features']['head_shoulders_8'] = hs3
+        p.uptick()
 
-    # get_trendlines(fund)
+        feature_plotter(fund, shapes, name=fund_name, feature='head_and_shoulders')
+        p.uptick()
+
+        filename = f"{fund_name}/candlestick_{fund_name}"
+        candlestick(fund, title=fund_name, filename=filename, saveFig=True)
+        p.uptick()
+
+        # get_trendlines(fund)
 
 
-# test_competitive(data, analysis)
+    # test_competitive(data, analysis)
 
-market_composite_index(period=period, properties=properties)
+    market_composite_index(config=config)
 
-bond_composite_index(period=period, properties=properties)
+    bond_composite_index(config=config)
 
-slide_creator(_DATE_REVISION_, analysis, _VERSION_)
-output_to_json(analysis)
+    slide_creator(_DATE_REVISION_, analysis, _VERSION_)
+    output_to_json(analysis)
 
-remove_temp_dir()
+    remove_temp_dir()
 
 print('Done.')
